PWM duty percent to float

+++ modules_test/ESC/lib/ESC_lib/ESC_lib.h
@@ -16,7 +16,7 @@ typedef struct {
         uint8_t pwm_channel  : 1;   /**< Canal PWM (0 o 1) dentro del slice. */
 
         uint16_t frequency_hz;      /**< Frecuencia PWM en Hz. */
-        uint8_t duty_percent;       /**< Pulso actual en microsegundos. */
+        float duty_percent;         /**< Porcentaje de duty cycle actual del PWM. */
         uint8_t speed_percent;       /**< Porcentaje de velocidad actual del motor. */
         uint16_t pwm_wrap;          /**< Valor máximo del contador PWM. */
         float clkdiv;               /**< Divisor de reloj para configurar frecuencia PWM. */
